Move advance_level to Champions app and add experience reward
